refactor: compute media details within component
